Refactoring of ADC reading on AAPS
Implemented general command IPC_CMD_GET_ADC
for both current and voltage measurements.
One byte that tells which type of measurement has
been done is trasfered in the response packet.

-Refactoring

+++ include/1wire.h
@@ -19,8 +19,8 @@ typedef struct {
 }ow_scratchpad_t;
 
 typedef struct {
-    uint16_t temp;
-    uint16_t dec;
+    uint8_t temp;
+    uint8_t dec;
 }ow_temp_t;
 ow_device_t *ow_devices;
 /* Long green  1-Wire D5 00 00 01 28 92 FD 28 */
